#colina-381-[DEPLOYED] - Inconsistency with Pagination (shared component pagination)
