+++ components/modal-content/labresultsview-modal-content.tsx
@@ -317,7 +317,6 @@ export const LabResultsViewModalContent = ({
 
         // Reset the selected file UUID
         setSelectedFileUUID("");
-
         // Close the delete modal
         setDeleteModalOpen(false);
       } else {
